requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.7
bcrypt==4.1.1
Pillow==11.1.0
reportlab==4.4.10
email-validator>=2.0

alembic==1.12.1
